Give expected results for negative non-base-10 numbers in tonumber().
This was undefined in Lua 5.1, but it's defined in 5.2.

(cherry picked from f3cf0d6e15240098147437fed7bd436ff55fdf8c)

`strtoul()` considers negative values as a valid input and silently
converts them to the equivalent unsigned long value. As a result yielded
value is unexpected to the user.

This patch adds reading of a sign (if exists) from argument and provide
the remaining part of the string as is if it starts with a digit or
alphabetical symbol to be consistent with Lua 5.2.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#6548