FFI/ARM64: Fix pass-by-value struct calling conventions.
(cherry picked from commit 9143e86498436892cb4316550be4d45b68a61224)

The arm64 parameters passing rules for Homogeneous Floating-point
Aggregates (HFA) are the following [1]:
* If the argument type is an HFA or an HVA, then the argument is used
  unmodified.
* If the argument is an HFA and there are sufficient unallocated
  Floating-point registers, then the argument is allocated to
  Floating-pointRegisters (with one register per member of the HFA).

Also, if the argument type is a Composite Type then the size of the
argument is rounded up to the nearest multiple of 8 bytes. LuaJIT FFI
backend makes this rounding unconditionally for arm64 architecture and
uses the result value to determine the necessary amount of registers
for the call. So for the HFA composed of 2n + 1 (< 4) float members the
extra one register is supposed as occupied. This leads to the wrong
value (0 due to the corresponding memset in `cconv_struct_init()`) in
this register if the other one parameter is passed to the procedure.

This patch fixes this case by using the real size of structure in the
calculation of necessary amount of registers to use.

Sergey Kaplun:
* added the description and the test for the problem

[1]: https://developer.arm.com/documentation/ihi0055/b/

Part of tarantool/tarantool#6548