jit: avoid Tarantool panic on os.exit
Using the `os.exit` function from the `bcsave.lua` module causes a
panic when called from the Tarantool because of an unexpected
fiber yield.

Since changing the `os.exit` implementation in the Tarantool sources
seems like a bad idea, the `bcsave.lua` module was changed. All of
the calls to `os.exit` were replaced with Lua errors with the same
error code.

Part of tarantool/tarantool#5541