tools: introduce parsers for sysprof
Since the sysprof's binary output is not human-readable, so there is a
demand to create a parser. The parser, which this commit provides,
converts sysprof's event stream into the format that the flamegraph.pl
can process.

The sysprof parser machinery uses the same symtab module as memprof's
parser since the format is the same.

Part of tarantool/tarantool#781