vm: save topframe info into global_State
Since commit 111d377d524e54e02187148a1832683291d620b2
('vm: introduce VM states for Lua and fast functions')
the VM has LFUNC and FFUNC states. The upcoming sampling
profiler uses these vmstates to determine if the guest
stack is valid or not. So, we need to provide a There is an inconsistent behavior
of the VM when the Lua stack is not valid, but the state
is set to LFUNC. This patch is just a gross hack with which
the profiler works fine. The problem is to be investigated
more deeply :(